--- commons/src/main/scala/za/co/absa/spline/common/WithResources.scala
@@ -1,59 +1,0 @@
-/*
- * Copyright 2019 ABSA Group Limited
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package za.co.absa.spline.common
-
-object WithResources {
-
-  import scala.util.control.NonFatal
-
-  def withResources[T <: AutoCloseable, V](r: => T)(f: T => V): V = {
-    val resource: T = r
-    require(resource != null, "resource is null")
-    var exception: Throwable = null
-    try {
-      f(resource)
-    } catch {
-      case e: Throwable =>
-        exception = e
-        throw e
-    } finally {
-      closeAndAddSuppressed(exception, resource)
-    }
-  }
-
-  private def closeAndAddSuppressed(e: Throwable, resource: AutoCloseable): Unit = {
-    if (e != null) {
-      try {
-        resource.close()
-      } catch {
-        case NonFatal(suppressed) =>
-          e.addSuppressed(suppressed)
-        case fatal: Throwable if NonFatal(e) =>
-          fatal.addSuppressed(e)
-          throw fatal
-        case fatal: InterruptedException =>
-          fatal.addSuppressed(e)
-          throw fatal
-        case fatal: Throwable =>
-          e.addSuppressed(fatal)
-      }
-    } else {
-      resource.close()
-    }
-  }
-
-}
