+++ parent/pom.xml
@@ -530,7 +530,7 @@
             <dependency>
                 <groupId>za.co.absa.commons</groupId>
                 <artifactId>commons_${scala.compat.version}</artifactId>
-                <version>0.0.6-SNAPSHOT</version>
+                <version>0.0.7</version>
             </dependency>
 
             <!-- Apache Commons -->
